backend/requirements.txt
Flask
Flask-CORS
langgraph
langchain-google-genai
python-dotenv
